Revert lint:infer-any warning when inferring Any/Nothing for HK
See discussions on PR 11157 / ticket 13128.
